Adding better file structure
