--- REFACTORING_SUMMARY.md
@@ -1,70 +1,0 @@
-# Task Management System Refactoring Summary
-
-## Overview
-
-This document summarizes the refactoring changes made to improve code organization, maintainability, and reusability in the Task Management System project.
-
-## Key Improvements
-
-### 1. UI Component Extraction
-
-Created reusable UI components in the `src/components/ui` directory:
-
-- **Button.jsx**: Flexible button component with variants, sizes, and states
-- **Input.jsx**: Form input component with consistent styling and error handling
-- **Select.jsx**: Dropdown select component with consistent styling
-- **Card.jsx**: Container component for content with consistent styling
-- **Modal.jsx**: Reusable modal dialog component
-- **StatusBadge.jsx**: Component for displaying task/project status with appropriate colors
-- **Message.jsx**: Component for displaying alerts, notifications, and feedback
-
-### 2. Layout Component Extraction
-
-Created layout components in the `src/components/layout` directory:
-
-- **AuthLayout.jsx**: Layout for authentication pages (sign in, sign up)
-- **DashboardLayout.jsx**: Layout for dashboard pages with sidebar, header, and content area
-
-### 3. Custom Hooks
-
-Created custom hooks in the `src/hooks` directory:
-
-- **useForm.js**: Hook for form state management and validation
-- **useLocalStorage.js**: Hook for working with localStorage
-- **useAuth.js**: Hook for authentication functionality
-
-### 4. Utility Functions
-
-Created utility functions in the `src/utils` directory:
-
-- **helpers.js**: Common utility functions for formatting, calculations, etc.
-
-### 5. Component Organization
-
-- Created index.js files in each directory for easier imports
-- Refactored page components to use the new layout and UI components
-- Improved component structure and separation of concerns
-
-### 6. Code Quality Improvements
-
-- Added JSDoc comments for better documentation
-- Improved naming conventions for better readability
-- Removed duplicate code by extracting common patterns
-- Enhanced component props with PropTypes for better type checking
-
-## Benefits
-
-1. **Improved Maintainability**: Smaller, focused components are easier to understand and maintain
-2. **Enhanced Reusability**: UI components can be reused across the application
-3. **Consistent Styling**: UI components ensure consistent styling throughout the application
-4. **Better Organization**: Clear directory structure makes it easier to find and work with components
-5. **Reduced Duplication**: Common patterns are extracted into reusable components and hooks
-6. **Better Developer Experience**: Improved imports, documentation, and component structure
-
-## Future Improvements
-
-1. Add more comprehensive form validation
-2. Implement a more robust state management solution
-3. Add unit tests for components and hooks
-4. Further optimize component rendering performance
-5. Implement more advanced UI features and animations
