+++ src/components/admin/index.js
@@ -1,0 +1,5 @@
+// Export all admin components for easier imports
+export { default as AdminHome } from './AdminHome';
+export { default as AdminProjects } from './AdminProjects';
+export { default as AdminTasks } from './AdminTasks';
+export { default as AdminChat } from './AdminChat';
