+++ src/components/student/index.js
@@ -1,0 +1,4 @@
+// Export all student components for easier imports
+export { default as StudentHome } from './StudentHome';
+export { default as StudentTask } from './StudentTask';
+export { default as StudentChat } from './StudentChat';
