--- src/utils/helpers.js
@@ -1,111 +1,0 @@
-/**
- * Generate a unique ID
- * @returns {string} A unique ID
- */
-export const generateId = () => {
-  return Date.now().toString(36) + Math.random().toString(36).substr(2, 5);
-};
-
-/**
- * Format a date string to a more readable format
- * @param {string} dateString - The date string to format
- * @param {Object} options - Formatting options
- * @returns {string} Formatted date string
- */
-export const formatDate = (dateString, options = {}) => {
-  try {
-    const date = new Date(dateString);
-    
-    const defaultOptions = {
-      year: 'numeric',
-      month: 'short',
-      day: 'numeric',
-      ...options
-    };
-    
-    return date.toLocaleDateString('en-US', defaultOptions);
-  } catch (error) {
-    console.error('Error formatting date:', error);
-    return dateString;
-  }
-};
-
-/**
- * Format status string for display
- * @param {string} status - The status string to format
- * @returns {string} Formatted status string
- */
-export const formatStatus = (status) => {
-  if (!status) return '';
-  
-  return status
-    .replace(/([A-Z])/g, ' $1')
-    .trim()
-    .split(' ')
-    .map(word => word.charAt(0).toUpperCase() + word.slice(1))
-    .join(' ');
-};
-
-/**
- * Calculate days remaining until a deadline
- * @param {string} dueDate - The due date
- * @returns {number} Days remaining
- */
-export const calculateDaysRemaining = (dueDate) => {
-  const today = new Date();
-  const due = new Date(dueDate);
-  
-  // Reset time part for accurate day calculation
-  today.setHours(0, 0, 0, 0);
-  due.setHours(0, 0, 0, 0);
-  
-  const diffTime = due - today;
-  const diffDays = Math.ceil(diffTime / (1000 * 60 * 60 * 24));
-  
-  return diffDays;
-};
-
-/**
- * Truncate text to a specified length
- * @param {string} text - The text to truncate
- * @param {number} maxLength - Maximum length before truncation
- * @returns {string} Truncated text
- */
-export const truncateText = (text, maxLength = 100) => {
-  if (!text || text.length <= maxLength) return text;
-  return text.slice(0, maxLength) + '...';
-};
-
-/**
- * Group array items by a specified key
- * @param {Array} array - The array to group
- * @param {string} key - The key to group by
- * @returns {Object} Grouped object
- */
-export const groupBy = (array, key) => {
-  return array.reduce((result, item) => {
-    const groupKey = item[key];
-    if (!result[groupKey]) {
-      result[groupKey] = [];
-    }
-    result[groupKey].push(item);
-    return result;
-  }, {});
-};
-
-/**
- * Sort array by a specified key
- * @param {Array} array - The array to sort
- * @param {string} key - The key to sort by
- * @param {string} direction - Sort direction ('asc' or 'desc')
- * @returns {Array} Sorted array
- */
-export const sortBy = (array, key, direction = 'asc') => {
-  const sortedArray = [...array];
-  
-  return sortedArray.sort((a, b) => {
-    if (a[key] < b[key]) return direction === 'asc' ? -1 : 1;
-    if (a[key] > b[key]) return direction === 'asc' ? 1 : -1;
-    return 0;
-  });
-};
